% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optsol_dGSMEF.R
\docType{methods}
\name{plot,optsol_dGSMEF,missing-method}
\alias{plot,optsol_dGSMEF,missing-method}
\title{plot}
\usage{
\S4method{plot}{optsol_dGSMEF,missing}(x, y, ylim = 50, xlab = "",
  ylab = "Value", type = "p", pch = 20, col = "black",
  plotRxns = NULL, plotFlux = NULL, baseline = 0, ...)
}
\description{
plot
}
\section{Methods (by class)}{
\itemize{
\item \code{x = optsol_dGSMEF,y = missing}: signature(x = "optsol_dGSMEF", y = missing"):
\itemize{
\item x: An object of class \code{optsol_dGSMEF}.  
\item y: not used but kept for compitability with generic plot.
\item plotRxns: List of reaction id's to be ploted.
}
}}

