% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optsol_dGSMEF.R
\docType{class}
\name{optsol_dGSMEF-class}
\alias{optsol_dGSMEF-class}
\title{Class "optsol_dGSMEF"}
\arguments{
\item{concmat}{Object of class "matrix" containing concentrations of extracellular metabolite and recombinant protein.}

\item{exRxn}{Object of class "matrix" containing names of exchange reactions for the EC metabolites.}

\item{tmVec}{Object of class "numeric" Vector of time points}

\item{bmVec}{Object of class "numeric" Vector of biomass concentrations}

\item{flux}{Object of class "fluxDistribution" containing the solutons flux distributions.}

\item{redCost}{Object of class "matrix" containing the solutons flux distributions.}

\item{Rxn}{Object of class "character" Vector of names of exchange reactions for the EC metabolites.}
}
\description{
Structure of class \code{optsol_dGSMEF}. Objects of that class are returned by the function \code{dGSMEF}.
}
\examples{
showClass("optsol_dGSMEF")
}
\seealso{
\code{dGSMEF}
}
